requirements.txt
cupy-cuda12x==13.4.0
glfw==2.8.0
imageio==2.37.0
imgui-bundle==1.6.2
loguru==0.7.3
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.1.0
plyfile==1.1
pyglm==2.8.0
PyOpenGL==3.1.4
PyYAML==6.0.2
setuptools
empy
catkin_pkg
lark
matplotlib
pandas